feat(build): TypeScript toolchain + typecheck gate (spec 06, Stage 1)
First step of the TypeScript migration (spec 06). Stands up the type
layer so .js and .ts coexist and gates regressions with `tsc --noEmit`
in both pre-commit and CI. No public .d.ts is emitted or shipped yet —
deferred to Stage 5, so we don't publish an incomplete (mostly-`any`)
surface over the community @types/three-nebula.

- Add `typescript` devDep and `tsconfig.json` (allowJs, checkJs off,
  strict off — ratcheted later; noEmit, esbuild does the transpile).
- Add `typecheck` script; run it in `ci:src` and in the husky pre-commit
  hook when any TS/JS or tsconfig is staged.
- Teach eslint-plugin-import to resolve `.ts` (import/resolver
  extensions) so .js files importing converted modules still lint.
- Convert the first leaf, `src/math/MathUtils.js` -> `.ts`, fully typed.

Stage 0 finding for spec 06: `core/three/Vector3.js` is a vendored
old-three Vector3 built with prototype assignment
(`Object.assign(Vector3.prototype, …)`) that TS can't infer, so
`Vector3D`'s inherited methods aren't visible on its type. MathUtils
bridges with a local structural interface until core/three is typed.
