+++ .eslintrc
@@ -33,5 +33,12 @@
     "import/named": 0,
     "newline-before-return": 1,
     "newline-after-var": 1
+  },
+  "settings": {
+    "import/resolver": {
+      "node": {
+        "extensions": [".js", ".ts"]
+      }
+    }
   }
 }
